pyproject.toml
[project]
name = "artigorioml"
version = "0.1.0"
description = "Códigos utilizados no estudo 'Técnicas de Aprendizado de Máquina para Predição de Gravidade de Acidentes em Rodovias do Estado do Rio de Janeiro'"
requires-python = ">=3.11.9"
dependencies = [
    "alembic==1.15.2",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "category-encoders==2.8.1",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "colorlog==6.9.0",
    "comm==0.2.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "daal==2025.5.0",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "fonttools==4.58.0",
    "fqdn==1.5.1",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "imbalanced-learn==0.13.0",
    "ipykernel==6.29.5",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "isoduration==20.11.0",
    "itables==2.3.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.0",
    "json5==0.12.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.6",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.5",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "kiwisolver==1.4.8",
    "lark==1.2.2",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.4.4",
    "notebook-shim==0.2.4",
    "numpy==2.2.5",
    "optuna==4.3.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "patsy==1.0.1",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyaml==25.1.0",
    "pyarrow==20.0.0",
    "pycparser==2.22",
    "pygeohash==3.1.3",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.3.0",
    "pytz==2025.2",
    "pywin32==310",
    "pywinpty==2.0.15",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rpds-py==0.26.0",
    "scikit-learn==1.6.1",
    "scikit-learn-intelex==2025.5.0",
    "scikit-optimize==0.10.2",
    "scipy==1.15.3",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sklearn-compat==0.1.3",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.40",
    "stack-data==0.6.3",
    "statsmodels==0.14.4",
    "tbb==2022.1.0",
    "tcmlib==1.3.0",
    "terminado==0.18.1",
    "threadpoolctl==3.6.0",
    "tinycss2==1.4.0",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20250708",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0"
]
